video_scribe/requirements.txt
requests
tqdm
modelscope
yt_dlp
openai
tenacity
json_repair
webvtt-py
